requirements.txt
rich
python-dotenv
openai
typer
requests
